Testcase_fix comment_ver1
fix 11 testcase follow comment .
