+++ Railway_Testing_Selenium/src/main/java/PageObjects/Railway/LoginPage.java
@@ -7,8 +7,8 @@ import org.openqa.selenium.WebElement;
 
 public class LoginPage extends GeneralPage{
     // Locators
-    private final By _txtUsername = By.xpath("//input[@id='username']");
-    private final By _txtPassword = By.xpath("//input[@id='password']");
+    private final By _txtUsername = By.id("username");
+    private final By _txtPassword = By.id("password");
     private final By _btnLogin = By.xpath("//input[@value='login']");
 
     // Elements
