+++ Railway_Testing_Selenium/src/test/java/Tsetcases/Railway/BaseTest.java
@@ -1,0 +1,20 @@
+package Tsetcases.Railway;
+
+import Common.Constant.Constant;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+public class BaseTest {
+    @BeforeMethod
+    public void beforeMethod() {
+        System.out.println("Pre-condition");
+        Constant.WEBDRIVER = new ChromeDriver();
+        Constant.WEBDRIVER.manage().window().maximize();
+    }
+    @AfterMethod
+    public void afterMethod() {
+        System.out.println("Post-condition");
+        Constant.WEBDRIVER.quit();
+    }
+}
