+++ Railway_Testing_Selenium/src/test/java/Tsetcases/Railway/ChangePasswordTest.java
@@ -4,24 +4,10 @@ import Common.Constant.Constant;
 import PageObjects.Railway.ChangePasswordPage;
 import PageObjects.Railway.HomePage;
 import PageObjects.Railway.LoginPage;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class ChangePasswordTest {
-    @BeforeMethod
-    public void beforeMethod() {
-        System.out.println("Pre-condition");
-        Constant.WEBDRIVER = new ChromeDriver();
-        Constant.WEBDRIVER.manage().window().maximize();
-    }
-    @AfterMethod
-    public void afterMethod() {
-        System.out.println("Post-condition");
-        Constant.WEBDRIVER.quit();
-    }
+public class ChangePasswordTest extends BaseTest {
     @Test
     public void TC09_UserCanChangePassword() {
         System.out.println("TC09 - User can change password");
