+++ Railway_Testing_Selenium/src/test/java/Tsetcases/Railway/RegisterTest.java
@@ -3,25 +3,10 @@ package Tsetcases.Railway;
 import Common.Constant.Constant;
 import PageObjects.Railway.HomePage;
 import PageObjects.Railway.RegisterPage;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class RegisterTest {
-    @BeforeMethod
-    public void beforeMethod() {
-        System.out.println("Pre-condition");
-        Constant.WEBDRIVER = new ChromeDriver();
-        Constant.WEBDRIVER.manage().window().maximize();
-    }
-    @AfterMethod
-    public void afterMethod() {
-        System.out.println("Post-condition");
-        Constant.WEBDRIVER.quit();
-    }
-
+public class RegisterTest extends BaseTest{
     @Test
     public void TC07() {
         System.out.println("TC07 - User can create new account");
@@ -64,11 +49,11 @@ public class RegisterTest {
         Assert.assertEquals(actualGeneralMsg, expectedGeneralMsg, "General error message is not displayed as expected");
 
         String actualPasswordMsg = registerPage.getValidPasswordMessage();
-        String expectedPasswordMsg = "Invalid password length";
+        String expectedPasswordMsg = "Invalid password length.";
         Assert.assertEquals(actualPasswordMsg, expectedPasswordMsg, "Password error message is not displayed as expected");
 
         String actualPidMsg = registerPage.getValidPidMessage();
-        String expectedPidMsg = "Invalid ID length";
+        String expectedPidMsg = "Invalid ID length.";
         Assert.assertEquals(actualPidMsg, expectedPidMsg, "PID error message is not displayed as expected");
     }
 }
